==> dev/8_explore.Rmd <==
---
title: "7_analysis"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(pacman)
p_load(tidyverse, scales, ggwordcloud)
```


1: SEE HOW SPEECHES WERE RATED BY MODEL

```{r compare metrics and speeches}
## original data
psp2010_tidy <- read_csv("data/csv/psp2010_tidy.csv") %>%
  select(-text_c)

l2010 <- list.files(path = "data/from_udpipe/", pattern = "_2010", full.names = T)

from_2010 <- map_df(l2010, read_csv) %>%
  rename(rowname = doc_id) %>%
  arrange(rowname) %>%
  filter(text_c != "0")

psp2010L <- full_join(psp2010_tidy, from_2010, by = "rowname")
rm(psp2010_tidy, l2010, from_2010)



## lda_output
comparison <- read_csv("data/csv/nov_tra_res_TEST.csv") %>%
  rename(rowname = doc_id) %>%
  left_join(psp2010L, by = "rowname")


```


```{r ploting}
ggplot(lda_output, aes(novelty, resonance)) +
    geom_bin2d(bins = 50) +
    scale_fill_viridis_c(trans = "log", breaks = c(1,10,100,1000),
                         labels = trans_format("log10", 
                                               math_format(expr = 10^.x, format = force))) +
    geom_smooth(method = "lm", colour = "black", linetype = "dashed", alpha = 0) +
    labs(x = "Novelty", y = "Resonance", 
         title = "Resonance vs. Novelty")

```


2: DIAGNOSE LDA TOPICS
```{r load model and hashes}
topic = read_delim("models/word_topic.model", delim = " ",
                   col_names = F, skip = 11,
                   col_types = cols()) %>%
  rename(hash = X1)

hash = read_csv("data/vw/hash_full.csv")

thr <- right_join(topic, hash, by = "hash")
```

```{r}
extract_topic_words <- function(column) {
  
  d <- thr %>%
    select(prob = column, word) %>%
    arrange(desc(prob)) %>%
    head(10) %>%
    mutate(topic = str_extract(column, "\\d+"),
           topic = as.numeric(topic)) %>%
    select(topic, word, prob)
  
  return(d)
  
}

col <- enframe(colnames(thr)) %>%
  filter(str_detect(value, "\\d+"))

col <- col$value

topic_words <- map_df(col, extract_topic_words)
```

```{r funky topic plots}
topic_words %>%
  filter(topic == "91") %>%
  ggplot(aes(label = word, size = prob / 2, color = factor(word))) +
  geom_text_wordcloud() +
  scale_size_area(max_size = 20) +
  theme_minimal()
```


27: administrative
40: regional administration
47: vysetrovani capaku
52: education and transportation
56: dostavba dukovan
70: dry season / climate change
84: appologies from non present MPs
91: budget
92: refugees
97: pauses
99: welfare state


# topic examples
```{r}
set.seed(420)

# capi hnizdo
topic_words %>%
  filter(topic == "47") %>%
  ggplot(aes(label = word, size = prob / 2, color = factor(word))) +
  geom_text_wordcloud() +
  scale_size_area(max_size = 20) +
  theme_minimal()

topic47 <- topic_words %>%
  filter(topic == "47") %>%
  mutate(word_en = c("dossier", "agrofert", "nest", "investigator", "stork", "nominee", "stork around", "spa", "come out", "occupy"))

topic47 %>%
  ggplot(aes(label = word_en, size = prob / 2, color = factor(word))) +
  geom_text_wordcloud() +
  scale_size_area(max_size = 20) +
  theme_minimal()
  
```

# lesy, klima, voda
```{r}
set.seed(420)

topic_words %>%
  filter(topic == "70") %>%
  ggplot(aes(label = word, size = prob / 2, color = factor(word))) +
  geom_text_wordcloud() +
  scale_size_area(max_size = 20) +
  theme_minimal()

topic70 <- topic_words %>%
  filter(topic == "70") %>%
  mutate(word_en = c("need", "react", "introduce", "water", "forest", "present", "forestry", "optionally", "definition", "leak"))

topic70 %>%
  ggplot(aes(label = word_en, size = prob / 2, color = factor(word))) +
  geom_text_wordcloud() +
  scale_size_area(max_size = 20) +
  theme_minimal()
```


# pauzy
```{r}
set.seed(420)

topic_words %>%
  filter(topic == "97") %>%
  ggplot(aes(label = word, size = prob / 2, color = factor(word))) +
  geom_text_wordcloud() +
  scale_size_area(max_size = 20) +
  theme_minimal()

topic97 <- topic_words %>%
  filter(topic == "97") %>%
  mutate(word_en = c("hour", "need", "minute", "a.k.a.", "pause", "meet", "punctual", "length", "technology", "declare"))

topic97 %>%
  ggplot(aes(label = word_en, size = prob / 2, color = factor(word))) +
  geom_text_wordcloud() +
  scale_size_area(max_size = 20) +
  theme_minimal()
```
